backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
alembic
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
